refactor: Start splitting Notification helpers in to use cases
Move enableNotifications and androidNotificationsAreEnabled to usecases.
