+++ app/src/main/java/app/pachli/MainActivity.kt
@@ -17,6 +17,7 @@
 
 package app.pachli
 
+import app.pachli.core.designsystem.R as DR
 import android.Manifest.permission.POST_NOTIFICATIONS
 import android.annotation.SuppressLint
 import android.app.NotificationManager
@@ -62,9 +63,9 @@ import app.pachli.appstore.EventHub
 import app.pachli.appstore.MainTabsChangedEvent
 import app.pachli.appstore.ProfileEditedEvent
 import app.pachli.components.compose.ComposeActivity.Companion.canHandleMimeType
-import app.pachli.components.notifications.androidNotificationsAreEnabled
 import app.pachli.components.notifications.createNotificationChannelsForAccount
-import app.pachli.components.notifications.enableAllNotifications
+import app.pachli.components.notifications.domain.AndroidNotificationsAreEnabledUseCase
+import app.pachli.components.notifications.domain.EnableAllNotificationsUseCase
 import app.pachli.core.activity.AccountSelectionListener
 import app.pachli.core.activity.BottomSheetActivity
 import app.pachli.core.activity.PostLookupFallbackBehavior
@@ -83,7 +84,6 @@ import app.pachli.core.data.repository.ListsRepository
 import app.pachli.core.data.repository.ListsRepository.Companion.compareByListTitle
 import app.pachli.core.database.model.AccountEntity
 import app.pachli.core.designsystem.EmbeddedFontFamily
-import app.pachli.core.designsystem.R as DR
 import app.pachli.core.model.Timeline
 import app.pachli.core.navigation.AboutActivityIntent
 import app.pachli.core.navigation.AccountActivityIntent
@@ -196,6 +196,12 @@ class MainActivity : BottomSheetActivity(), ActionButtonActivity, MenuProvider {
 
     @Inject
     lateinit var developerToolsUseCase: DeveloperToolsUseCase
+
+    @Inject
+    lateinit var enableAllNotifications: EnableAllNotificationsUseCase
+
+    @Inject
+    lateinit var androidNotificationsAreEnabled: AndroidNotificationsAreEnabledUseCase
 
     private val binding by viewBinding(ActivityMainBinding::inflate)
 
@@ -1069,8 +1075,8 @@ class MainActivity : BottomSheetActivity(), ActionButtonActivity, MenuProvider {
 
         // Setup notifications
         // TODO: Continue to call this, as it sets properties in NotificationConfig
-        androidNotificationsAreEnabled(this, accountManager)
-        lifecycleScope.launch { enableAllNotifications(this@MainActivity, mastodonApi, accountManager) }
+        androidNotificationsAreEnabled(this)
+        lifecycleScope.launch { enableAllNotifications(this@MainActivity) }
 
         updateProfiles()
